Refactor WinRMConfig and use it in VMWare builder
In order to allow the build pipeline to modify the WinRM configuration
it's necessary to pass around a reference to the actual WinRMConfig
object rather than extracting the username and password while
constructing the pipeline itself.

To achieve this we also make the VMWare builder use the common
WinRMConfig structure rather than the identical but internal one it was
previously using.