Add HTTPS options to the WinRMConfig and propagate
This commit adds members to the WinRM configuration struct for HTTPS
(defaults to true) and ignoring the certificate chain (useful if you
have self-signed certificates). It also adds options to the
configuration for providing a custom CA certificate to validate against,
though this is not currently propagated through to the communicator.

It also changes the default port if none is specified to 8956 from 8955,
reflecting the change to HTTPs by default.