Add auto-configuration of WinRM over HTTPS
This commit introduces a new step into the build process which generates
a user-data powershell script responsible for configuring WinRM access
over HTTPS. The script does the following:

- Disables the firewall rules for WinRM over HTTP for both public and
  domain/private profiles
- Sets the maximum timeout for WinRM to 1800000ms
- Sets the maximum memory per WinRM shell to 1024MB
- Disallows unencrypted WinRM connections
- Enables Basic authentication for WinRM
- Enables CredSSP access to WinRM (necessary for using it internally)
- Adds firewall rules for port 5986 to both public and domain/private
  profiles
- Adds the given certificate to the LocalMachine\My certificate store
- Configures a listener for WinRM on HTTPS using the specified
  certificate
- Restarts the WinRM service

Note that the certificate is sent base-64 encoded in the user-data and
is therefore visible via the console for the period of time that the
builder is running. Future work may want to involve creating an S3
bucket and an IAM role into which the instance is provisioned so this is
not as accessible, though it's instance specific.

The certificate and WinRM configuration are currently left installed. It
would be a nice option to be able to remove them after configuration but
that's not straightforward because scheduling a shutdown script in
Windows is hard. In this case it would also be better to add options for
a certificate chain to validate against, and the option to generate the
certificate chain internally in Packer rather than relying on OpenSSL or
mkcert/New-SelfSignedCertificate as we currently do.

As part of this change it was necessary to move to a local custom
version of StepRunInstance as the version in Packer receives all it's
configuration during the pipeline build phase instead and doesn't allow
modification. Not sure if there was an original design reason for that
but this version seems somewhat neater anyway and involves a lot less
copying around of values.