Add support for changing Administrator password
It may be desirable for users to specify a new Administrator password
rather than retrieving the generated one from Amazon. To acheive this we
reset the local Administrator password to a known value as part of the
user data script which configures WinRM.

Consequently this can only be used with the winrm_autoconfigure option
turned on, though that isn't necessarily a long-term limitation.