Set POSH ExecutionPolicy to Bypass in provisioners
This commit changes from setting the Powershell ExecutionPolicy to
RemoteSigned as part of the user data script to setting it to Bypass by
default as part of the ExecuteCommand for provisioners. This seems
neater as we don't change the machine more than necessary underneath the
user and remove the need to override the ExecuteCommand every time a
provisioner is used.